project/requirements.txt
aiortc==1.6.0
aiohttp==3.9.1
websockets==12.0
opencv-python==4.8.1.78
numpy==1.24.3
cryptography==41.0.7
pycryptodome==3.19.0
tkinter-modern==1.0.0
Pillow==10.1.0
asyncio-mqtt==0.16.1
python-socketio==5.10.0
flask==3.0.0
flask-socketio==5.3.6
eventlet==0.33.3